requirements.txt
absl-py==0.9.0
arff2pandas==1.0.1
ase==3.21.1
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
backports.entry-points-selectable==1.1.1
black==21.11b1
bleach==3.1.1
blis==0.4.1
boto==2.49.0
boto3==1.14.7
botocore==1.17.7
cachetools==4.1.0
catalogue==1.0.0
catalyst==19.3rc0
certifi==2019.11.28
cffi==1.13.2
cfgv==3.3.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
ConfigArgParse==1.2.3
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.3
Cython==0.29.16
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.3
docutils==0.15.2
easydict==1.10
en-core-web-sm==2.2.5
entrypoints==0.3
filelock==3.4.0
Flask==1.1.2
FlyAI==0.7.1
future==0.18.2
geoopt==0.2.0
gevent==20.6.2
google-auth==1.17.0
google-auth-oauthlib==0.4.1
googledrivedownloader==0.4
greenlet==0.4.16
grpcio==1.29.0
gym==0.17.2
h5py==2.10.0
hyperlink==21.0.0
hypernymysuite==0.0.1
identify==2.4.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
incremental==21.3.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.16.0
jieba==0.42.1
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
jpeg4py==0.1.4
jsonpatch==1.26
jsonpointer==2.0
jsonrpc==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
liac-arff==2.5.0
lltm==0.0.0
llvmlite==0.35.0
lmdb==1.4.0
Markdown==3.2.2
MarkupSafe==1.1.1
mat4py==0.4.3
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.12.0
mpmath==1.1.0
murmurhash==1.0.2
mypy-extensions==0.4.3
natsort==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nodeenv==1.6.0
notebook==6.0.3
numba==0.52.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
pandas==1.3.4
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
platformdirs==2.4.0
plotly==5.3.1
pre-commit==2.16.0
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.8.0
ptranking==0.0.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycparser==2.19
pyglet==1.5.0
Pygments==2.5.2
pynlp==0.4.2
pyonmttok==1.18.4
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyrouge==0.1.3
pyrsistent==0.15.7
python-dateutil==2.8.1
python-louvain==0.15
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qrcode==6.1
qtconsole==4.6.0
rdflib==5.0.0
redis==3.5.3
regex==2021.11.10
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.1
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentencepiece==0.1.85
service-identity==18.1.0
six==1.13.0
sklearn==0.0
smart-open==2.0.0
spacy==2.2.3
srsly==1.0.1
stanford-openie==1.0.3
stanfordcorenlp==3.9.1.1
stanfordnlp==0.2.0
subword-nmt==0.3.7
sympy==1.6
tenacity==8.0.1
tensorboard==2.2.2
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
toml==0.10.2
tomli==1.2.2
torch==1.5.0
torch-cluster==1.5.7
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.7
torch-spline-conv==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchtext==0.4.0
torchvision==0.2.1
tornado==6.0.3
tqdm==4.30.0
traitlets==4.3.3
Twisted==21.2.0
typed-ast==1.5.1
typing-extensions==4.0.1
ufile==3.2.2
urllib3==1.25.8
virtualenv==20.10.0
visdom==0.1.8.9
visual-genome==1.0.0
waitress==1.4.4
wasabi==0.6.0
watermark==2.2.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
zipp==3.0.0
zope.event==4.4
zope.interface==5.1.0
